finagle-core: Move TransportClientConnection Out of StdStackServer
Problem / Solution

It would be beneficial for testing if we could pick and choose pieces
of a server and stitch them together without having to build a full on
`StackServer`. This can be done for a client today, but not for a server.
Some necessary functionality is contained directly within the logic for
`StdStackServer`, including the ability to create a
`TransportClientConnection`. Let's extract `TransportClientConnection` so
that it can be used outside.

Differential Revision: https://phabricator.twitter.biz/D325375