finagle-mysql: Use PlainHandshake Directly in ClientDispatcher
Problem

For `finagle-mysql`, handshaking within the `ClientDispatcher` is
considered a legacy path, and will be eventually removed. So, we don't
want to add any new handshaking functionality to that pathway. Therefore,
we should be instantiating `PlainHandshake` directly instead of calling
`Handshake.apply` so that the `ClientDispatcher` is only ever capable of
instantiating plain text sessions.

Solution

Change the use of `Handshake.apply` to `new PlainHandshake` in the
`ClientDispatcher`.

Differential Revision: https://phabricator.twitter.biz/D326705