finagle-mysql: Remove ClientSsl from Transporter Params
Problem

If a `finagle-mysql` client desires to use SSL/TLS, a `Transport.ClientSsl`
param will be passed to the `MysqlTransporter`. However, that param cannot
be passed directly to the underlying `Netty4Transporter`, because doing so
will cause SSL/TLS to attempt to be negotiated as part of connection
establishment, instead of waiting until the appropriate point in handshaking.

Solution

Remove the `Transport.ClientSsl` param from the collection of params that are
sent to the underlying `Netty4Transpoter` so that the `SslHandler` is not
added to the Netty pipeline initially.

Result

The `Netty4Transporter` does not attempt to negotiate SSL/TLS as part of
connection establishment, but we still can add it at the right time later.

Differential Revision: https://phabricator.twitter.biz/D326780