finagle-mysql: Don't log exceptions if we fail to rollback a transaction and it was a ChannelClosedException
Problem:
If a rollback fails due to a ChannelClosedException, the logs are very
spammy and leads devs triaging issues to falsely blame them.

Solution:
If the rollback fails due to a ChannelClosedException (which happens if the
underlying service has already been closed by a lower-level call), don't log
the exception

Result:
Any rollbacks that fail due to a ChannelClosedExceptions don't log the exception,
but we still poison the connection and try and close it again.

JIRA Issues: CSL-8242

Differential Revision: https://phabricator.twitter.biz/D327111