finagle/scrooge: Generate less classes
Problem

While merging Scrooge PR 310, it was noted that a few things could be
tightened up.

Solution

Remove Scrooge's versions of `ThriftMethodStats` which were generated
per service and instead use the version in
`com.twitter.finagle.thrift`.  Additionally, create a singleton
``ThriftMethodStats.Null` which can be used as a sentinel.

Differential Revision: https://phabricator.twitter.biz/D329764