finagle-core: ensure ClientDispatcher queueSize gauge is removed on transport close
Problem

The `GenSerialClientDispatcher` has a `queueSize` gauge
that gets registered for every instance. The `HttpClientDispatcher`
extends from this class, thus registering this gauge.
For the HTTP/2 implementation, a `HttpClientDispatcher`
is created for every new stream, increasing the number of
registered gauges. As currently implemented, these gauges
will be cleaned up upon GC.

Solution

Be a little more aggressive and remove the registered
gauge on the close of the transport, instead of waiting
for GC.

Result

Less memory pressure from gauges and less overhead
during metrics collection.

JIRA Issues: CSL-8378

Differential Revision: https://phabricator.twitter.biz/D331923