finagle-core: Remove SslClientEngineFactory.getHostname
Problem

The use of `SslClientEngineFactory.getHostname` is highly discouraged
because it uses the `getHostName` method of the passed in `InetSocketAddress`
if the hostname within the `SslClientConfiguration` is None. The `getHostName`
method of `InetSocketAddress` performs a reverse DNS lookup when provided with
an IP address, which is undesriable.

Solution

Remove `SslClientEngineFactory.getHostname` and encourage folks to use
`SslClientEngineFactory.getHostString` instead.

JIRA Issues: CSL-8416

Differential Revision: https://phabricator.twitter.biz/D334087