finagle-base-http: Add support for new "b3" tracing header
Problem

OpenZipkin has created a new set of header to carry
trace information. These headers are not compatible
with the old ones and are needed to interoperate
with services that use the new version of the header.

Solution

If a "b3" header exists in the request this extracts
all the new information and populates finagles
tracing header with the existing values.

https://github.com/twitter/finagle/pull/749
Signed-off-by: Ian Bennett <ibennett@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D334419