finagle-core: Allow to use SOCKS proxy for localhost
Problem

SOCKS proxy is bypassed when connecting to localhost and there's no way to
override this behaviour. Using a SOCKS proxy when connecting to localhost can be
useful to easily capture communication between local services for debugging.

Solution

Add a flags to enable SOCKS proxy for localhost.

Result

You can now force the use of SOCKS proxy for localhost with
-com.twitter.finagle.socks.socksProxyForLocalhost. Without this flag, the
behaviour of bypassing the SOCKS proxy is kept.

Signed-off-by: Moses Nakamura <mnakamura@twitter.com>

JIRA Issues: CSL-8450

Differential Revision: https://phabricator.twitter.biz/D337073