finagle-core: OffloadFilter flag to reduce network contention
**Problem**

We simulated a scenario when the service is close to its redline and profiled it with vtune. One of the top methods is a lock queue used for networking (`queued_spin_lock_slowpath`)

**Solution**

The contention seemed to be due to finagle applications reusing the netty event loop to run application code. Vladimir mentioned the new `OffloadFilter` that was created to deal with applications that do blocking operations and we decided to re-use it to test the hypothesis. The results are promising. After the change to use the offload filter with a fixed thread pool by default, there is a significant reduction in contention

**Notes**

- Looking at the [[ https://t.co/Ls5YZYpX2n?amp=1 | source code ]] of the system call, we noticed that the contention seems to happen when there are concurrent flushes and that the lock happens at the level of the network device. Our theory is that having fewer threads allocated to the event loop reduces the concurrency and makes the mux batching more effective, reducing the rate of flushes. It's possible that applications running with this configuration will become better neighbors when running in a shared aurora host since the lock happens at the network device level.

- Additionally, with this new configuration the netty threads are always available to handle channel events, helping to reduce the latency. Since a channel is always pinned to a specific thread by netty, without the offload filter channels eventually have to wait for the thread to finish executing the application code to have their network events handled.

- I haven't created any tests since the stack parameter is initialized statically. Please let me know if you think of a way to unit test the change.

Differential Revision: https://phabricator.twitter.biz/D331502