finagle-http: Move the creation of the endpointer out of the Http object
Problem

The code in the Http.scala file is getting pretty hairy, making it tough
to make changes that may benefit HTTP/2.

Solution

Change the http client implementation configuration to be based on
providing an endpointer instead of transporter oriented tooling and
put this all in a new object, ClientEndpointer.

Differential Revision: https://phabricator.twitter.biz/D337136