finagle: modify build.sbt file to allow for testing with Netty SNAPSHOT build
Problem

We have a manual process of validating Netty versions post-release. This
results in discovering breakages late in the cycle and can delay moving
to the next release.

Solution

Modify the `build.sbt` file to allow for providing properties to change
the Netty build version. If the build is a SNAPSHOT build, the snapshot
repo will be added to the resolvers.

GitOrigin-RevId: e6580fe653dc93b4ee1e901bacb7ca6e87c064fe