finagle: modify build.sbt file to allow for testing with Netty SNAPSHOT build
Problem

We have a manual process of validating Netty versions post-release. This
results in discovering breakages late in the cycle and can delay moving
to the next release.

Solution

Modify the `build.sbt` file to allow for providing properties to change
the Netty build version. If the build is a SNAPSHOT build, the snapshot
repo will be added to the resolvers.

JIRA Issues: CSL-8566

Differential Revision: https://phabricator.twitter.biz/D346056