finagle-core: Improve concurrency in NackAdmissionFilter
Problem

Under heavy throughput clients see lock contention in the `Ema` used
by `NackAdmissionFilter`.

Solution

Replace the implementation with a version that trades some loss of
accuracy for less contention.

This tradeoff seems worthwhile, as the assumption is that for updates
that come in at roughly the same time do not have a meaningful effect
on the moving average. This is a result of the math used in the
computation. More details are available in `LossyEma.update`.

Result

No more contention.

GitOrigin-RevId: 1732fd18099105aeed93388b66b5c632d01b55b5