finagle-http2: Swich multiplex codec based clients to a better pooling strategy
Problem

We don't have the correct representation of the H2 session in our stack.
Specifically, we treat each stream as a unique session which is
wasteful.

Solution

Develop a dynamic pooling strategy that switches between the singleton
pool or the default pool for the H2 and H1 sessions, respectively.

JIRA Issues: CSL-8437

Differential Revision: https://phabricator.twitter.biz/D346236