finagle-http2: The MultiplexCodec backed ClientSession should close when streams are exhausted
Problem

We don't currently know what the behavior will be when we exhaust the
stream identifiers available for H2 (the maximum unsigned 31bit
integer).

Solution

Be proactive and declare a H2 session closed once the number of
remaining stream identifiers is less than 50. This will cause the
pooling layer to discard the session.

GitOrigin-RevId: ff1c9e5cf53aaed3f0da04050a7f89c7c323b0ab