finagle-core: Improve concurrency in NackAdmissionFilter
Problem

Under heavy throughput clients see lock contention in the `Ema` used
by `NackAdmissionFilter`.

Solution

Replace the implementation with a version that trades some loss of
accuracy for less contention.

This tradeoff seems worthwhile, as the assumption is that for updates
that come in at roughly the same time do not have a meaningful effect
on the moving average. This is a result of the math used in the
computation. More details are available in `LossyEma.update`.

This fixes the previous attempt at this in ``D346532``. That was
reverted as it was missing the update to `lastTime` in
`LossyEma.update`.  This patch takes care of that and adds more tests.

Result

No more contention.

JIRA Issues: CSL-8614

Differential Revision: https://phabricator.twitter.biz/D351249