finagle-http2: fix location of H2UriValidatorHandler in pipeline
Problem

The `H2UriValidatorHandler` is essentially dead code, as its
placement in the Netty pipeline results in never seeing
HTTP/2 frames. We were still returning a 400 response, because
the post HTTP/2 -> HTTP/1.1 conversion would enter the
`UriValidatorhandler` before being removed from the pipeline.

Solution

Put the `H2UriValidatorHandler` in a place where it can see
the HTTP/2 frames earlier. Fix the type of response we're
returning and update the tests.

Result

We will eject bad URI paths in a pure HTTP/2 pipeline earlier
and with less overhead.

GitOrigin-RevId: da239c1c0b318935d8c9947b6e27e0388dacd21f