finagle-http2: add Ping FailureDetector to Multiplex based clients
Problem/Solution

MultiplexCodec based HTTP/2 clients do not have support for
Ping based failure detection. Let's add it and have pings
handle status in a similar style to Mux.

Note: The FailureDetector.Param is still set to NullConfig
and is not modified as part of this change. Pings will
not be enabled until that is changed.

JIRA Issues: CSL-8250

Differential Revision: https://phabricator.twitter.biz/D352720