finagle-core: Remove synchronization from CachingPool
Problem

High throughput clients see lock contention in `CachingPool`.

Solution

Change the data structure used into one that does not need
synchronization.

The prior implementation in ``D339657`` had 2 race conditions.  One
was between `pollFirst` and `removeIf`, where `removeIf` could end up
closing a `Service` that was already in use. This was addressed by
adding additional state to the `WrappedService` so that usage would be
exclusive.

The other existed previously, but was less likely to be triggered and
would happen when a transport level read timeout would fall around the
same time as the CachingPool's expiry. This was addressed by running
the expiry task more frequently.

Result

Less contention.

GitOrigin-RevId: c8988b67d14a7651e3f2344ec4bbc07b81ad683b