finagle-http2: Fix starvation condition when H2 fails to upgrade
Problem

When we abort the H2C upgrade we switch the Netty pipeline back to the
configured backpressure, typically auto-read off. This may result in the
channel no longer thinking there is a read interest even though the
`ChannelTransport` is awaiting a message, leading to starvation.

Solution

Make sure we signal interest in reading at least one more message.

GitOrigin-RevId: c39d3e8f8714b182b7e4abc01a2e12c93a3e9665