finagle-http: ensure streaming tests test HTTP/2 properly
Problem

The `AbstractStreamingTest` exposes an `impl` property
to allow clients and servers to swap out implementation.
The `impl` property is never used when configuring the
clients, meaning that HTTP/2 is not being tested. We
also do not have explicit coverage for the legacy and
`MultiplexHandler` based HTTP/2 clients.

Solution

Modify the way clients and servers get configured for
the tests so that we can remove the `impl` property.
Be more explicit about whether or not to use HTTP/2.

Result

We now test HTTP/2 clients for streaming.

JIRA Issues: CSL-8769

Differential Revision: https://phabricator.twitter.biz/D363420