finagle-http: Add 'None' as valid SameSite value
Problem

The SameSite attribute is intended to accept Lax, Strict and None as values (0). Currently, only Lax and Strict is supported.

Also, as motivation, Chrome is supposedly planning to treat SameSite=None as a way to signal you want to use x-site cookies as of February 2020 (1).

Solution

Added None in the sealed trait as a valid case.

Result

You can set SameSite=None in the cookie.

(0): https://tools.ietf.org/html/draft-west-first-party-cookies-07
(1): https://blog.chromium.org/2019/05/improving-privacy-and-security-on-web.html

Signed-off-by: Moses Nakamura <mnakamura@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D365170