finagle-core: Rearranged BUILD files to avoid relative paths
Problem

Pants doesn't play nicely when you have ".." in your path, because it makes it
harder to track changes to your file system.  Nice and easy when you just have
to look in subdirectories.

Solution

Move the BUILD file up a level, and add a redirect at the old location.

JIRA Issues: DPB-11645

Differential Revision: https://phabricator.twitter.biz/D368349