finagle-redis: Add hashed-key based partitioning/routing support to redis
Problem:

It would be advantageous in some circumstances to be able to use a cluster of
redis servers without requiring the use of a central coordinating service.

Solution:

Take advantage of the new finagle-partitioning module to implement a
partitioned client for redis.

Result:

A new redis client stack that uses the (somewhat ill-named)
KetamaPartitioningService to provide key-based routing to a number of backend
servers. The standard and transactional clients have not changed. This review
only adds a new configuration, and handles an appropriate subset of redis
commands (ones that do not have an implicit assumption that they're talking to
a single backend node). Note that the hashing algorithm is configurable, and
the code is hash-implementation agnostic.

GitOrigin-RevId: e56c6ceef599b1886b6cca192a99e208280d2a87