finagle-core: better FailureAccrual logging
Problem
The failure accrual module logs only the endpoints that are marked dead.
It is hard to understand from the logs which policy was violated and for how
long the endpoint is marked dead.

Solution
Make `FailureAccrualPolicy` a `Showable`. The policy's state is included in
the logs

JIRA Issues: CSL-8732

Differential Revision: https://phabricator.twitter.biz/D369209