finagle-mux: Propagate SupersededRequestFailure Multiple Levels
Problem

When using 'BackupRequests' via the `BackupRequestFilter`, generally
retrieval of information from multiple levels of downstream services
is necessary. Of the two requests, the `BackupRequestFilter` will raise
a `SupersededRequestFailure` on the loser (if the winner is successful).
The Mux server properly handles the receipt of the `SupersededRequestFailure`
on the first level, however it doesn't properly propagate it beyond its own
level. This means that requests cancelled multiple levels deep because of a
`SupersededRequestFailure` are not as `Ignorable` only `Interrupted`. This has
a negative effect on the success rate metrics of downstream services.

Solution

Properly handle multiple level `SupersededRequestFailure` propagation so that
exceptions are raised which are considered `Ignorable`.

Result

The success rate of downstream services should be more accurate in regards to
ignorable failures from backup requests.

JIRA Issues: CSL-8780

Differential Revision: https://phabricator.twitter.biz/D365729