finagle+scrooge: Pass a factory to create TReusableBuffer as the parameter of a finagle client
Problem

Currently every instance of finagle client holds its own TReusableBuffer. The purpose of this object
is to hold an instance of temporary buffer for every io thread. A typical size of the buffer is around
16k. This means that if, for example, there are 128 i/o threads then one client uses at least two megabytes of heap.
If the app creates multiple instances of the client then the memory allocated for these buffers is just wasted
because the thread can use only one buffer at any given time and the number of threads is limited.

Solution
Pass a shared instance of TReusableBuffer to the client's ctor.

Result
Significantly decreased memory usage

GitOrigin-RevId: 62ccebaee3632c3ff24e5f030216125a5be62ca2