finagle-zipkin-core: Reduce default value for initialSampleRate flag
Problem

Every currently untraced RPC is subject to a sampling decision at each server,
controlled by the flag initialSampleRate. The default is currently 0.1%, which
is exceptionally high for tracing deployed at scale (e.g. at Twitter). Because
requests tend to fan out towards the leaves, this leads to an imbalance between
full traces (i.e. those with roots at ingress) and partial traces (with roots
lower down the dependency graph).

For example, if a request has depth=10, fanout=2, then we expect to have 2^10 =
1024 leaf nodes per root node, so we're 1024 times as likely to start a trace at
a leaf node as at a root node. In combination with a high default sampling rate,
this leads to the trace corpus being dominated by single-span traces, which
significantly reduces the number of traces suitable for aggregate and end-to-end
analysis.

Solution

Add the capability to dynamically configure the sampling rate by moving it
from Sampler class instances to the Sampler companion object. This does not
change the ability of service owners to override the default setting using the
-initialSampleRate flag.

Result

If desired, the default sampling rate can be modified without requiring every
service to pass a command-line flag, and without changing the existing
default value of 0.1%.

JIRA Issues: CSL-8814

Differential Revision: https://phabricator.twitter.biz/D374634