oss: update all projects with JDK 11 support and travis builds
Problem

We want to be able to build and test using JDK 11, but
Dodo and Scrooge do not use compatible Scala and SBT
versions.

Solution

Update Scala and SBT versions used by dodo and scrooge
from Scala 2.10.6 to 2.10.7 and SBT from 0.13.16 to
0.13.18. Each SBT file will contain GC options that
are compatible with the active JDK version. Some logic
has been updated to fix runtime behavior in JDK 11.

Result

Projects can be compiled and tested by both JDK 8 and JDK 11.

GitOrigin-RevId: 5a7456420520fdeb7cf164345f1f5a6775571b41