finagle-core: aperture Ring.range precision
Problem
When the width is 1.0, we expect the range of indices to span the entire vector
set. Due to a precision error, this equality `(offset + 1.0) % 1.0 == offset`
does not hold true, causing `Ring.range` to return 0, when `offset == (n/offset)`
for n > 0.

Solution / Result
Modify `Ring.range` to detect when modulo operation is unnecessary when wrapping
around the ring.

GitOrigin-RevId: 22a6d0df4b78621204f5b526301a478f8730f841