finagle-http2: Don't use RstExceptions for nack responses
Problem

We currently use a RstException for all RST frames. This can get
surfaced to users in metrics which can be tricky to understand. It also
means we can distinguish from all the other causes of RST which are much
more interesting.

Solution

In the case of a nack RST, synthesize a 503 nack response that will get
converted in the HttpNackFilter in the exact same way that it does for
HTTP/1.x.

GitOrigin-RevId: e8df6b5b24711844b3378642766d9e5057b63132