finagle-http: Fix singleton pool bug in Http2StreamingTest
Problem

Some of the tests in Http2StreamingTest are expecting that there is only
a single, serial session available to the application. However, this
isn't true for the prior-knowledge HTTP/2 implementation which uses the
singleton pool making the singular session shareable.

Solution

In the case where we want to constrain the pool size to 1 we switch back
to the default pool module and limit the size to 1, just like in
HTTP/1.x.

JIRA Issues: CSL-9016

Differential Revision: https://phabricator.twitter.biz/D392399