finagle-core: fixed aperture status check
Problem:
Checking the status of the Aperture balancer will loop through the
entire vector set even though clients only communicate with endpoints within the
aperture. This will lead the a consistent Status.Open balancer even if all the
nodes within the aperture are Status.Busy

`Ring.range` was not safe on the hot-path due to a precision error fixed in
D388243.

Solution/Result:
Limit the status check to nodes within the aperture.

RandomAperture still considers the entire vector set when computing it's status.
This is due to RandomAperture having cases in which `needsRebuild == true`. Any
circuit breaker residing above the Balancer that make use `status` as a signal
could cause RandomAperture to never to rebuilds due rebuilds occuring in
`Balancer.apply`

JIRA Issues: CSL-8856

Differential Revision: https://phabricator.twitter.biz/D390258