finagle-base-http: Make `path` Public for `Uri`
Problem / Solution

finagle-base-http's `Uri` class is capable of parsing the full URI
into a path and query parameters. Currently only the params are publicly
accessible. In code in Twitter Server, providing access to both the params
and the path would allow us to remove duplicated code. Let's provide access
to the path publicly.

GitOrigin-RevId: e3fd2de12fbcb9933660b9e66c2e5577b995af72