finagle-http2: Simplify UpgradeSuccess
Problem

We have some legacy indirection that was necessary to support the older
H2 implementation, but that is now gone, and we have unnecessary
complexity.

Solution

Clean things up by making UpgradeSuccess simply carry the expected values
instead of a function that ignores its only argument and returns the
expected values.

Cleanup how we set the H2 session

GitOrigin-RevId: 0b7c4bd149aa28a207e7cff7052b0a97638b519e