util: Add a tool to make metrics lazy
Problem

Sometimes we don't know in advance if we're going to be using certain metrics.
However, we don't want to have to change our programming model to have to know
in advance if things are lazy.

Solution

Provide a LazyStatsReceiver that will automatically make metrics lazy without
having to rewrite where they're used.

GitOrigin-RevId: 6199b1994138ab1c8e1d99076be7ea27c482e3bb