finagle-thrift: Added back in a couple of constructors because of diamond deps
Problem

If you depend on stale java generated code as well as tip of Finagle, you can
run into a sticky situation where you try to new up a RichClientParam but it
doesn't exist anymore.

Solution

Add back in a couple of the RCP constructors that are used in Java generated
code at least until we publish a new Finagle + Scrooge.

NO-QUEUE=true
GitOrigin-RevId: 603c0df5ab4517546bc39fd6078786f5b0e3addb