finagle-core,util-core,util-benchmark: tag traces with microsecond timestamp resolution
Problem

Currently traces are stamped with millisecond resolution timestamps. This makes
it hard to sequence within traces because events often occur less than a
millisecond apart. In addition we cannot derive meaningful information about
relative timing of events.

Solution/Result

Offer a new method on the Time object to allow us to gather timestamps at
nanosecond resolution so that, when using a JDK that supports it, we will
have more precise timing information in traces.

GitOrigin-RevId: c8342e29e249fee8e2a35ecf00116f68b3e3ab28