finagle doc: documentation around newService/newClient
Problem:
Finagle client documenation explains newService and refers to the newClient
api for stateful clients. No additional documenation is provided beyond this.

Solution/Result:
Promote newService to the top of the finagle Clients page, as it's the most
commonly used api. Added documentation in regards to how sessions are
established with `newClient` and closing sessions to avoid resource leaks.

GitOrigin-RevId: 8223e59a51dd33562b48f54f6c6a7fef2ea38aa4