finagle doc: documentation around newService/newClient
Problem:
Finagle client documenation explains newService and refers to the newClient
api for stateful clients. No additional documenation is provided beyond this.

Solution/Result:
Promote newService to the top of the finagle Clients page, as it's the most
commonly used api. Added documentation in regards to how sessions are
established with `newClient` and closing sessions to avoid resource leaks.

JIRA Issues: CSL-9113

Differential Revision: https://phabricator.twitter.biz/D406898