finagle-netty4: Use ref-counted SSL contexts and engines
Problem

We spend a lot of GC time due to finalizers on the Netty4 generated
SslEngine instances.

Solution

Use the ref-counted API. We don't have a mechanism for easily managing
the lifetime of the parent context so we wrap it in a proxy that itself
has a finalizer and simply forwards all function calls to the parent.

Differential Revision: https://phabricator.twitter.biz/D377828