finagle-stats: Faster Glob matching
Problem

Verbosity levels can be adjusted via a glob expression passed as a flag or a tunable. We evaluate it
on every request to `/metrics.json` and want to do better than converting to a more powerful
regex construct.

Solution

Introduce `Glob` matcher, which is at least 10x faster than what we're doing before.

GitOrigin-RevId: 4c175fa86658bc44cef30c2f95efe143b92bd467