finagle-core: Consider SSLExceptions with the string 'SSLEngine closed already' ChannelClosedExceptions
Problem

Having a Netty SslHandler in the pipeline turns what was originally a
ChannelClosedException to a SslException because all writes are
processed by the SslHandler which fails because its SSLEngine is already
closed.

Solution

Since these exceptions are essentially the same thing as a
ChannelClosedException, lets just translate them to that.

JIRA Issues: CSL-8988

Differential Revision: https://phabricator.twitter.biz/D408123