finagle-core: Consider SSLExceptions with the string 'SSLEngine closed already' ChannelClosedExceptions
Problem

Having a Netty SslHandler in the pipeline turns what was originally a
ChannelClosedException to a SslException because all writes are
processed by the SslHandler which fails because its SSLEngine is already
closed.

Solution

Since these exceptions are essentially the same thing as a
ChannelClosedException, lets just translate them to that.

GitOrigin-RevId: 5308d8da8ff721ffc1a6a8c8831e14a634ab6f5a