finagle-serversets: Ensure ZkSession#retrying is resilient to ZK host resolution failure
**Problem**
Currently, there is no exception handling in `ZkSession#retrying`,
since we assume `newZkSession` will always return normally.  However, it's
possible that `newZkSession` throws `UnknownHostException` if DNS fails and we
cannot resolve hosts.

This causes the thread responsible for re-establishing the `ZkSession` to
terminate and the corresponding `ServiceDiscoverer` to remain unhealthy forever.

**Solution**
To solve this, we add a new `WatchState` to indicate that a
`ZkSession` could not be created (`WatchState.FailedToInitialize`). Then we
adapt the reconnect logic in `ZkSession#retrying` to retry on
`WatchState.FailedToInitialize` (in addition to session expiry).

**Result**
`ZkSession#retrying` can recover from intermittent DNS failures.

GitOrigin-RevId: d3223566ea9b10335270ebee63230a68afcdb500