DarkTrafficFilter annotations
Problem

We cannot tag individual client spans emanating from the DarkTrafficFilter
because we do not have access to the tracing context contained within their
stacks. This makes it impossible to annotate dark traffic spans as well as being
able to identify which light and dark spans are tied together logically. This
information is important in allowing us to understand how request paths are
structured and what pieces are important when analyzing overall performance.

Solution/Result

Place a module in the client stack to allow us to capture context information
forwarded by the DarkTrafficFilter. Use this information within the correct
tracing context to add a binary annotation to the span. Dark requests will
contain a "clnt/dark_request" and a "clnt/dark_request_key" annotation and the
corresponding light request will contain the exact same "clnt/dark_request_key"
so that they can be directly tied together.

GitOrigin-RevId: 4156a70df6e4f0d438b985c36090434243dbee5f