finagle-partitioning: Fix build dependencies to not run finagle-core tests
Problem

The finagle-partitioning test package is dependent on
"finagle/finagle-core/src/test/scala" meaning it will run all the tests
in finagle-core as well. This dependency exists to bring in the
StringServer and similar types which are useful for building basic server
and client instances.

Solution

Change the dependency to
"finagle/finagle-core/src/test/scala:stringprotocol" which brings in
only the testing utils.

JIRA Issues: CSL-9236

Differential Revision: https://phabricator.twitter.biz/D415307