finagle-http2: Http2 server cleans up its pipeline
Problem

We don't clean up the tail of our HTTP/2 socket pipelines on the server
side and that can leave some things dangling around which can be bad for
GC and cause unexpected behavior, such as undesirable timeouts.

Solution

Clean up the tail of our pipeline when we convert over to HTTP/2.

JIRA Issues: CSL-9341

Differential Revision: https://phabricator.twitter.biz/D429416