finagle-http2: Change close behavior of H2 servers
Problem

When we convert a HTTP/1.x pipeline to a HTTP/2 pipeline we are still at
the mercy of the tail ChannelTransport which can trigger a close in the
case of a session timeout.

Solution

Swallow the close events once we convert to a HTTP/2 server pipeline and
redirect the Server.close calls to the H2ServerFilter which will take
charge of doing the graceful shutdown for servers.

JIRA Issues: CSL-9348

Differential Revision: https://phabricator.twitter.biz/D429554