finatra/http: Add toBufReader into StreamingResponse
Problem
`StreamingResponse` only carries Json streams, valid Json arrays.
However, sometimes there are cases we want to add prefixes or suffixes
of the Json array.

Solution
Add a `toBufReader` to get the serialized `Reader[Buf]`, it can be used to
be modified to any form as long as they are Bufs.
Add a `Reader.concat` to merge a collection of Readers to a single Reader
for convenience.

GitOrigin-RevId: b6d228a1075faad354205dcd703707ecc0d666ab