finagle-core: Remove default NullTracer from ClientBuilder
Problem

A number of production services still use the ClientBuilder API, but do not
override the default NullTracer. This limits the usefulness of distributed
tracing for debugging and performance analysis.

Solution

Remove the default NullTracer. Although ClientBuilder is deprecated, this will
improve the quality of traces in the interim period until the relevant services
migrate to Finagle 6 APIs.

Result

Traces will have more complete client-side instrumentation.

GitOrigin-RevId: 798ad48b522c5a71a35cc6ab3cbcc05b3e7f8fc5