finagle-http2: Change frame logging to INFO level
Problem

The HTTP/2 frame logging code logs to it's various logger instances at level
TRACE which is not typically enabled. This means that when you enable the
frame logger you all but certainly have to change the individual logger's levels
to make their entries surface. This makes debugging harder than it needs to be
and is inconsistent with the level of INFO used in the ChannelSnooper code
which plays the same role, just for Netty channel messages.

Solution

Change the logging level to INFO.

GitOrigin-RevId: a534e2038eacf23b07a7e7d3a377d20769670da2