util|scrooge|finagle: Update to Scala 2.13.1
Problem/Solution

Update util, scrooge, and finagle to all cross build
with Scala 2.13.1. We need to also update the
`scoverage` plugin here to deal with an issue where
plugins need to be specified to be fully cross versioned
(that is published with their full scala version, e.g., `2.13.0`
instead of just `2.13`) because Scala is not binary compatible
between minor releases and something changed between
2.13.0 and 2.13.1. The `scoverage` plugin version `1.6.1`
appears to work properly.

Differential Revision: https://phabricator.twitter.biz/D444211