finagle-partitioning: Rename CacheNode to PartitionNode
Problem

CacheNode doesn't represent itself as it is not cache specifically. It
represents any partition node that carries its address and metadata.

Solution

Rename CacheNode and CacheNodeMetadata to PartitionNode and
PartitionNodeMetadata.

Result

Along with rename, also added `fromAddress` and `fromInetSocketAddress` (private
in finagle now) to let specify a node without host and port.

GitOrigin-RevId: 088453423e0c9ef11662c6ff7e56118f9504877e