finagle-thriftmux: Remove ThriftMux Client to Thrift Server Test
Problem / Solution

A ThriftMux client connected to a Thrift server causes a Thrift server
to allocate excessive memory, leading to the Thrift server potentially
receiving an out of memory error and closing the connection. In previous
versions of pants testing facilities, it was ok for this failure
scenario to occur, but not anymore. Therefore we remove this specific
test as we know that a ThriftMux client connecting to a Thrift server is
bad news.

JIRA Issues: CSL-9539

Differential Revision: https://phabricator.twitter.biz/D450981