finagle-thrift: Add a thrift request serializer
Problem

For messaging fan-out, we need to serialize sub-requests from the original
request and then send each to wire routing to their partitions.

Solution

Add a ThriftRequestSerializer, this is very similar to the scrooge generated
`encodeRequest`, this doesn't need to be context.Local since we can access to
the ThriftStruct.write and other necessary parameters.

Result

Add access to TReusableBufferFactory in client stack.

Differential Revision: https://phabricator.twitter.biz/D452763