finagle-netty4: ChannelStatsHandler shouldn't count multiple close calls
Problem

We can get racy calls to `.close(..)` from a `ChannelTransport` or
similar constructs that bridge the finagle and Netty worlds. This will
result in multiple increments to the `closed` metric, although only a
single channel will be closed.

Solution

Like we do with `channelInactive` stats, we guard the `closed` metric
behind a flag.

JIRA Issues: CSL-9615

Differential Revision: https://phabricator.twitter.biz/D462360